# PR #3: Comprehensive Testing Framework for Unified Dashboard System