… tests/performance/test_dashboard_performance.py
… tests/test_comprehensive_dashboard_integration.py
… tests/test_dashboard.py